Refactor quantization into a single struct with a shared data store.
Binary and scalar quantization share a struct with an internal enum
managing the type of quantization, including serde logic. This allows us
to straighten out the `quantize` command, and accessing quantized vector
data can be done through a single store that knows how to quantize a
query.

With this change it should be easier to build an n-bit binary quantizer
that may using hamming distance scoring.