Implement Statistical Binary Quantization (SBQ).
Compute mean and standard deviation of a sample of the vector store and
use this to quantize into an n-bit unary coded value that can be
compared using hamming distance.

On my 1536d data set using a 2-bit value is worth 3 additional points of
recall, and 3-bits is worth an additional 3 points. This is without any
quantized re-ranking which may help further.

Refactor facilities around quantization and reading quantized stores to
make things easier to work with when considering all the options.