Implement generic dequantization and dequantized rescoring.
Add this functionality to the Quantizer attached to the store and
implement a scorer that will dequantize vectors to perform euclidean
scoring.

This helps for sbq but not quite as much as I hoped. Not sure if it has
to do with the score distribution produced by SBQ or if there's just too
little data compared to int4 quantization.