Implement scalar quantization confidence bounds.
This is implemented in the same way as lucene: sample vectors and
eliminate outliers in the 1/dim quantiles of each sample, then compute
the average min and max. Average min and max are stored in the codec and
values are quantized into fixed size buckets with outliers going at
either end of the range.

This does not integrate well with existing build and search bits, we may
need a QuantizedEuclideanScorer or similar that handles all formats.

In the long run we may wish to use tdigest or similar to compute the
quantiles without buffering so much.