Greatly speed scoring for packed scalar vectors.
Load greater amounts at once: load 128 bits in all cases which unpacks
to 64 elements (2-bit) or 32 elements (4-bit).

Unpack the query just once when using a QueryScorer instead of unpacking
for each score() call. This is only applied in the search path but could
easily help in the build path.