Introduce sharding for building and searching the graph.
The idea behind doing this is that lucene implicitly shards the graph by
time so might as well see what this is like.

This does not require any changes to the immutable graph format, which
is not very good for such a sparse graph. When we search a sharded graph
we use the same num_candidates setting for all graphs, then merge the
result sets when done. Vectors are i7 quantized so we do not bother with
re-ranking.

shards	num_candidates	recall@10	recall@100	avg latency
1	100		0.94474		0.89825		 3.222 ms
10	100		0.97159		0.97186		15.977 ms
1	1000		0.98378		0.98339		11.036 ms

Interestingly enough, using the same effective budget on a larger graph
yields both lower latency and higher recall.